Remove and move some things, also callb ack a function.
