+++ src/app/loadout/ModPicker.tsx
@@ -14,14 +14,13 @@ import { RootState } from 'app/store/types';
 import { DestinyClass, DestinyEnergyType, DestinyProfileResponse } from 'bungie-api-ts/destiny2';
 import raidModPlugCategoryHashes from 'data/d2/raid-mod-plug-category-hashes.json';
 import _ from 'lodash';
-import React from 'react';
+import React, { useCallback } from 'react';
 import { connect } from 'react-redux';
 import { createSelector } from 'reselect';
-import '../../item-picker/ItemPicker.scss';
-import { isLoadoutBuilderItem } from '../item-utils';
-import { knownModPlugCategoryHashes, slotSpecificPlugCategoryHashes } from '../known-values';
-import { isInsertableArmor2Mod, sortMods } from '../mod-utils';
-import PlugDrawer from './PlugDrawer';
+import { isLoadoutBuilderItem } from './item-utils';
+import { knownModPlugCategoryHashes, slotSpecificPlugCategoryHashes } from './known-values';
+import { isInsertableArmor2Mod, sortMods } from './mod-utils';
+import PlugDrawer from './plug-drawer/PlugDrawer';
 
 /** Slot specific mods can have at most 2 mods. */
 const MAX_SLOT_SPECIFIC_MODS = 2;
@@ -140,54 +139,54 @@ function mapStateToProps() {
  * A sheet to pick mods that are required in the final loadout sets.
  */
 function ModPicker({ mods, language, lockedMods, initialQuery, onAccept, onClose }: Props) {
-  const isModSelectable = (
-    mod: PluggableInventoryItemDefinition,
-    selected: PluggableInventoryItemDefinition[]
-  ) => {
-    const { plugCategoryHash } = mod.plug;
-    const isSlotSpecificCategory = slotSpecificPlugCategoryHashes.includes(plugCategoryHash);
+  const isModSelectable = useCallback(
+    (mod: PluggableInventoryItemDefinition, selected: PluggableInventoryItemDefinition[]) => {
+      const { plugCategoryHash } = mod.plug;
+      const isSlotSpecificCategory = slotSpecificPlugCategoryHashes.includes(plugCategoryHash);
 
-    let associatedLockedMods: PluggableInventoryItemDefinition[] = [];
+      let associatedLockedMods: PluggableInventoryItemDefinition[] = [];
 
-    if (isSlotSpecificCategory || plugCategoryHash === armor2PlugCategoryHashesByName.general) {
-      associatedLockedMods = selected.filter(
-        (mod) => mod.plug.plugCategoryHash === plugCategoryHash
-      );
-    } else if (raidModPlugCategoryHashes.includes(plugCategoryHash)) {
-      associatedLockedMods = selected.filter((mod) =>
-        raidModPlugCategoryHashes.includes(mod.plug.plugCategoryHash)
-      );
-    } else {
-      associatedLockedMods = selected.filter(
-        (mod) => !knownModPlugCategoryHashes.includes(mod.plug.plugCategoryHash)
-      );
-    }
+      if (isSlotSpecificCategory || plugCategoryHash === armor2PlugCategoryHashesByName.general) {
+        associatedLockedMods = selected.filter(
+          (mod) => mod.plug.plugCategoryHash === plugCategoryHash
+        );
+      } else if (raidModPlugCategoryHashes.includes(plugCategoryHash)) {
+        associatedLockedMods = selected.filter((mod) =>
+          raidModPlugCategoryHashes.includes(mod.plug.plugCategoryHash)
+        );
+      } else {
+        associatedLockedMods = selected.filter(
+          (mod) => !knownModPlugCategoryHashes.includes(mod.plug.plugCategoryHash)
+        );
+      }
 
-    // We only care about this for slot specific mods and it is used in isModSelectable. It is calculated here
-    // so it is only done once per render.
-    const lockedModCost = isSlotSpecificCategory
-      ? _.sumBy(associatedLockedMods, (mod) => mod.plug.energyCost?.energyCost || 0)
-      : 0;
+      // We only care about this for slot specific mods and it is used in isModSelectable. It is calculated here
+      // so it is only done once per render.
+      const lockedModCost = isSlotSpecificCategory
+        ? _.sumBy(associatedLockedMods, (mod) => mod.plug.energyCost?.energyCost || 0)
+        : 0;
 
-    if (isSlotSpecificCategory) {
-      // Traction has no energy type so its basically Any energy and 0 cost
-      const modCost = mod.plug.energyCost?.energyCost || 0;
-      const modEnergyType = mod.plug.energyCost?.energyType || DestinyEnergyType.Any;
+      if (isSlotSpecificCategory) {
+        // Traction has no energy type so its basically Any energy and 0 cost
+        const modCost = mod.plug.energyCost?.energyCost || 0;
+        const modEnergyType = mod.plug.energyCost?.energyType || DestinyEnergyType.Any;
 
-      return (
-        associatedLockedMods.length < MAX_SLOT_SPECIFIC_MODS &&
-        lockedModCost + modCost <= MAX_ARMOR_ENERGY_CAPACITY &&
-        (modEnergyType === DestinyEnergyType.Any || // Any energy works with everything
-          associatedLockedMods.some((l) => l.plug.energyCost?.energyType === modEnergyType) || // Matches some other enery
-          associatedLockedMods.every(
-            (l) =>
-              (l.plug.energyCost?.energyType || DestinyEnergyType.Any) === DestinyEnergyType.Any
-          )) // If every thing else is Any we are good
-      );
-    } else {
-      return associatedLockedMods.length < MAX_SLOT_INDEPENDENT_MODS;
-    }
-  };
+        return (
+          associatedLockedMods.length < MAX_SLOT_SPECIFIC_MODS &&
+          lockedModCost + modCost <= MAX_ARMOR_ENERGY_CAPACITY &&
+          (modEnergyType === DestinyEnergyType.Any || // Any energy works with everything
+            associatedLockedMods.some((l) => l.plug.energyCost?.energyType === modEnergyType) || // Matches some other enery
+            associatedLockedMods.every(
+              (l) =>
+                (l.plug.energyCost?.energyType || DestinyEnergyType.Any) === DestinyEnergyType.Any
+            )) // If every thing else is Any we are good
+        );
+      } else {
+        return associatedLockedMods.length < MAX_SLOT_INDEPENDENT_MODS;
+      }
+    },
+    []
+  );
 
   return (
     <PlugDrawer
