+++ src/app/loadout/plug-drawer/Footer.m.scss
