+++ src/app/loadout/plug-drawer/Footer.tsx
