+++ src/app/loadout/plug-drawer/PlugDrawer.tsx
