+++ src/app/loadout/plug-drawer/PlugSection.m.scss
