+++ src/app/loadout/plug-drawer/PlugSection.tsx
